renamed GranularityValidator
