--- src/main/java/com/rackspace/ceres/app/config/configValidator/iGranularityValidator.java
@@ -1,25 +1,0 @@
-package com.rackspace.ceres.app.config.configValidator;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-import javax.validation.Constraint;
-import javax.validation.Payload;
-
-@Constraint(validatedBy = GranularityValidator.class)
-@Target({ ElementType.TYPE })
-@Retention(RetentionPolicy.RUNTIME)
-public @interface iGranularityValidator {
-
-  String message() default "Granularities are not multiples of each other";
-  Class<?>[] groups() default {};
-  Class<? extends Payload>[] payload() default {};
-  /*
-  @Constraint(validatedBy = FieldsValueMatchValidator.class)
-@Target({ ElementType.TYPE })
-@Retention(RetentionPolicy.RUNTIME)
-public @interface FieldsValueMatch {
-   */
-
-}
